code/instance/code/requirements.txt
tqdm==4.64.0
firebase-admin==5.2.0
google-cloud-compute==1.4.0